requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
brave-search==0.1.8
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cssselect==1.2.0
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
InstructorEmbedding==1.0.1
jaraco.classes==3.4.0
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.1
langchain-core==0.1.45
langsmith==0.1.49
llama-index==0.10.30
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.12
llama-index-core==0.10.30
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-instructor==0.1.3
llama-index-embeddings-openai==0.1.8
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.9
llama-parse==0.4.1
llamaindex-py-client==0.1.18
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
minijinja==1.0.16
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.23.2
orjson==3.10.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
playwright==1.43.0
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
pypdf==4.2.0
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.8.1
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.19.0
sentence-transformers==2.7.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
tomlkit==0.12.4
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
trio==0.25.0
trio-websocket==0.11.1
trove-classifiers==2024.4.10
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
